Remove app controller
